Move runjs into tests directory
Most of the code in this file relates to running of tests and its not
part of the toolchain itself.
