+++ tools/ctor_evaller.py
@@ -19,7 +19,7 @@ import time
 
 sys.path.insert(1, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from tools import shared, js_optimizer, jsrun, building
+from tools import shared, js_optimizer, building
 from tools.tempfiles import try_delete
 
 
@@ -268,7 +268,7 @@ console.log(JSON.stringify([numSuccessful, Array.prototype.slice.call(heap.subar
     err_file_handle = open(err_file, 'w')
     proc = subprocess.Popen(shared.NODE_JS + [temp_file], stdout=out_file_handle, stderr=err_file_handle, universal_newlines=True)
     try:
-      jsrun.timeout_run(proc, timeout=10, full_output=True, throw_on_failure=False)
+      shared.timeout_run(proc, timeout=10, full_output=True, throw_on_failure=False)
     except Exception as e:
       if 'Timed out' not in str(e):
         raise
@@ -315,7 +315,7 @@ def eval_ctors_wasm(js, wasm_file, num):
   logger.debug('wasm ctor cmd: ' + str(cmd))
   proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
   try:
-    err = jsrun.timeout_run(proc, timeout=10, full_output=True, throw_on_failure=False)
+    err = shared.timeout_run(proc, timeout=10, full_output=True, throw_on_failure=False)
   except Exception as e:
     if 'Timed out' not in str(e):
       raise
